Depuracion

+++ Laboratorio3ED2/Compresor/Estructuras/NodoHuff.cs
@@ -5,8 +5,8 @@ namespace Compresor.Huffman
     public class NodoHuff<T> : IComparable where T : IComparable
     {
         public T Value { get; set; }
-        public int Frecuencia { get; set; }
-        public decimal ProbPrio { get; set; }
+        public decimal Frecuencia { get; set; }
+        public int ProbPrio { get; set; }
         public NodoHuff<T> Derecho { get; set; }
         public NodoHuff<T> Inzquierdo { get; set; }
